# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QPX Quantum Progress Experience API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PlayerProfile {
  id: number;
  username: string;
  /** Unique display ID like QPX-1042 */
  playerId: string;
  /** @nullable */
  avatarUrl?: string | null;
  level: number;
  /** Current XP in this level */
  xp: number;
  xpToNextLevel: number;
  totalXp: number;
  hp: number;
  maxHp: number;
  coins: number;
  /** Current daily streak in days */
  streak: number;
  longestStreak?: number;
  /** Current rank name: Recruit, Cadet, Trainee, Warrior... */
  rank: string;
  /** 0-based rank index (0=Recruit ... 9=Titan) */
  rankIndex: number;
  /**
     * Equipped legendary title
     * @nullable
     */
  title?: string | null;
  onboardingComplete: boolean;
  /** @nullable */
  primaryGoal?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface PlayerProfileUpdate {
  username?: string;
  avatarUrl?: string;
  bio?: string;
  title?: string;
}

export interface OnboardingInput {
  username: string;
  age: number;
  height: string;
  weight: string;
  primaryGoal: string;
  topPriorities: string[];
  skillsToLearn: string[];
  dailyHours: number;
  biggestWeakness: string;
  habitsToLearn?: string[];
  /**
     * unused - keep for compat
     * @nullable
     */
  habitsToLearn_extra?: string | null;
  habitsToBuild: string[];
  desiredIdentity: string;
}

export interface XpAwardInput {
  amount: number;
  reason: string;
}

export interface XpAwardResult {
  player: PlayerProfile;
  xpGained: number;
  leveledUp: boolean;
  newLevel: number;
  newRank: string;
  rankPromoted: boolean;
}

export interface Quest {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** mind | body | knowledge | creation | discipline | health | productivity */
  category: string;
  xpReward: number;
  coinReward: number;
  /** HP lost if quest is missed */
  hpPenalty?: number;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  /** YYYY-MM-DD date this quest belongs to */
  date: string;
  createdAt: string;
}

export interface QuestInput {
  title: string;
  description?: string;
  category: string;
  xpReward: number;
  coinReward?: number;
  hpPenalty?: number;
  date?: string;
}

export interface QuestUpdate {
  title?: string;
  description?: string;
  category?: string;
  xpReward?: number;
  coinReward?: number;
  hpPenalty?: number;
}

export interface Achievement {
  id: number;
  title: string;
  description: string;
  category: string;
  /** @nullable */
  icon?: string | null;
  xpReward: number;
  coinReward: number;
  unlocked: boolean;
  /** @nullable */
  unlockedAt?: string | null;
  /** common | rare | epic | legendary */
  rarity: string;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  target?: number | null;
}

export interface QuestCompleteResult {
  quest: Quest;
  player: PlayerProfile;
  xpGained: number;
  coinsGained: number;
  leveledUp: boolean;
  rankPromoted: boolean;
  newAchievements?: Achievement[];
}

export interface JournalEntry {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  wins?: string | null;
  /** @nullable */
  lessons?: string | null;
  /** @nullable */
  goals?: string | null;
  /** excellent | good | neutral | low | terrible */
  mood: string;
  /** 1-10 */
  energyLevel: number;
  createdAt: string;
  updatedAt?: string;
}

export interface JournalEntryInput {
  title: string;
  content: string;
  wins?: string;
  lessons?: string;
  goals?: string;
  mood: string;
  energyLevel: number;
}

export interface JournalEntryUpdate {
  title?: string;
  content?: string;
  wins?: string;
  lessons?: string;
  goals?: string;
  mood?: string;
  energyLevel?: number;
}

export interface CategoryStat {
  category: string;
  count: number;
  /** @nullable */
  xp?: number | null;
}

export interface RankHistoryEntry {
  rank: string;
  achievedAt: string;
}

export interface PlayerStats {
  totalXp: number;
  totalCoins: number;
  totalQuestsCompleted: number;
  currentStreak: number;
  longestStreak: number;
  totalJournalEntries: number;
  totalAchievementsUnlocked: number;
  questsByCategory: CategoryStat[];
  xpByCategory: CategoryStat[];
  rankHistory: RankHistoryEntry[];
}

export interface ProgressDataPoint {
  date: string;
  xp: number;
  level: number;
  questsCompleted?: number;
}

export interface BossBattle {
  id: number;
  title: string;
  description: string;
  category?: string;
  targetValue: number;
  currentProgress: number;
  xpReward: number;
  coinReward: number;
  /** @nullable */
  badgeReward?: string | null;
  /** active | completed | failed | upcoming */
  status: string;
  expiresAt: string;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface BossBattleProgressInput {
  increment: number;
}

export interface InventoryItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** outfit | effect | frame | badge | title */
  type: string;
  /** common | rare | epic | legendary */
  rarity: string;
  unlocked: boolean;
  equipped: boolean;
  /** @nullable */
  requiredRank?: string | null;
  /** @nullable */
  requiredLevel?: number | null;
}

export interface Inventory {
  items: InventoryItem[];
  /** @nullable */
  equippedOutfit: string | null;
  /** @nullable */
  equippedEffect: string | null;
  /** @nullable */
  equippedFrame: string | null;
}

export interface LoginReward {
  /** coins | xp | badge_fragment | special | rare */
  type: string;
  amount: number;
  label: string;
}

export interface DailyLoginResult {
  alreadyClaimed: boolean;
  reward: LoginReward;
  streakDay: number;
  player: PlayerProfile;
}

export interface CalendarDay {
  day: number;
  reward: LoginReward;
  claimed?: boolean;
  special?: boolean;
}

export interface DailyLoginCalendar {
  rewards: CalendarDay[];
  claimedDays: number[];
  currentStreak: number;
}

export interface Season {
  id: number;
  name: string;
  subtitle: string;
  startDate: string;
  endDate: string;
  active: boolean;
  /** @nullable */
  description?: string | null;
  rewards?: string[];
}

export type GetQuestsParams = {
/**
 * Date in YYYY-MM-DD format, defaults to today
 */
date?: string;
};

export type GetProgressHistoryParams = {
/**
 * Number of days of history (default 30)
 */
days?: number;
};