# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QPX Quantum Progress Experience API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quests:
    get:
      operationId: getQuests
      tags: [quests]
      summary: List all quests for today
      parameters:
        - name: date
          in: query
          schema:
            type: string
          description: Date in YYYY-MM-DD format, defaults to today
      responses:
        "200":
          description: List of quests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quest"
    post:
      operationId: createQuest
      tags: [quests]
      summary: Create a new quest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestInput"
      responses:
        "201":
          description: Created quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
  /quests/{id}:
    patch:
      operationId: updateQuest
      tags: [quests]
      summary: Update a quest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestUpdate"
      responses:
        "200":
          description: Updated quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
    delete:
      operationId: deleteQuest
      tags: [quests]
      summary: Delete a quest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quests/{id}/complete:
    post:
      operationId: completeQuest
      tags: [quests]
      summary: Mark a quest as complete and award XP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quest completed with XP result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestCompleteResult"

  # ── Achievements ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PlayerProfile:
      type: object
      required: [id, username, playerId, level, xp, xpToNextLevel, hp, maxHp, coins, streak, rank, rankIndex, totalXp, onboardingComplete, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        playerId:
          type: string
          description: Unique display ID like QPX-1042
        avatarUrl:
          type: ["string", "null"]
        level:
          type: integer
        xp:
          type: integer
          description: Current XP in this level
        xpToNextLevel:
          type: integer
        totalXp:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        coins:
          type: integer
        streak:
          type: integer
          description: Current daily streak in days
        longestStreak:
          type: integer
        rank:
          type: string
          description: "Current rank name: Recruit, Cadet, Trainee, Warrior..."
        rankIndex:
          type: integer
          description: 0-based rank index (0=Recruit ... 9=Titan)
        title:
          type: ["string", "null"]
          description: Equipped legendary title
        onboardingComplete:
          type: boolean
        primaryGoal:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Quest:
      type: object
      required: [id, title, category, xpReward, coinReward, completed, date, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          description: "mind | body | knowledge | creation | discipline | health | productivity"
        xpReward:
          type: integer
        coinReward:
          type: integer
        hpPenalty:
          type: integer
          description: HP lost if quest is missed
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
        date:
          type: string
          description: YYYY-MM-DD date this quest belongs to
        createdAt:
          type: string
    QuestInput:
      type: object
      required: [title, category, xpReward]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        xpReward:
          type: integer
        coinReward:
          type: integer
        hpPenalty:
          type: integer
        date:
          type: string
    QuestUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        xpReward:
          type: integer
        coinReward:
          type: integer
        hpPenalty:
          type: integer
    QuestCompleteResult:
      type: object
      required: [quest, player, xpGained, coinsGained, leveledUp, rankPromoted]
      properties:
        quest:
          $ref: "#/components/schemas/Quest"
        player:
          $ref: "#/components/schemas/PlayerProfile"
        xpGained:
          type: integer
        coinsGained:
          type: integer
        leveledUp:
          type: boolean
        rankPromoted:
          type: boolean
        newAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"

    # ── Achievements ───────────────────────────────────────────────────────────
    Achievement:
      type: object
      required: [id, title, description, category, xpReward, coinReward, unlocked, rarity]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: ["string", "null"]
        xpReward:
          type: integer
        coinReward:
          type: integer
        unlocked:
          type: boolean
        unlockedAt:
          type: ["string", "null"]
        rarity:
          type: string
          description: "common | rare | epic | legendary"
        progress:
          type: ["integer", "null"]
        target:
          type: ["integer", "null"]

    # ── Journal ────────────────────────────────────────────────────────────────
...