# lib/api-zod/src/generated/types/questCompleteResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QPX Quantum Progress Experience API
 * OpenAPI spec version: 0.1.0
 */
import type { Achievement } from './achievement';
import type { PlayerProfile } from './playerProfile';
import type { Quest } from './quest';

export interface QuestCompleteResult {
  quest: Quest;
  player: PlayerProfile;
  xpGained: number;
  coinsGained: number;
  leveledUp: boolean;
  rankPromoted: boolean;
  newAchievements?: Achievement[];
}