# lib/api-zod/src/generated/types/questInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QPX Quantum Progress Experience API
 * OpenAPI spec version: 0.1.0
 */

export interface QuestInput {
  title: string;
  description?: string;
  category: string;
  xpReward: number;
  coinReward?: number;
  hpPenalty?: number;
  date?: string;
}